集成 laravel- permission 扩展包

+++ database/migrations/2025_03_11_155807_add_parent_id_to_replies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('replies', function (Blueprint $table) {
-            $table->unsignedBigInteger('parent_id')->nullable()->after('id'); // 允许为空
+            $table->unsignedInteger('parent_id')->nullable()->after('id'); // 允许为空
             $table->foreign('parent_id')->references('id')->on('replies')->onDelete('cascade'); // 级联删除
         });
     }
